add page

+++ data/src/main/java/com/badcompany/data/model/CarColorEntity.kt
@@ -3,7 +3,7 @@ package com.badcompany.data.model
 /**
  * Representation for a [CarColorEntity] fetched from an external layer data source
  */
-data class CarColorEntity(val id: Int,
+data class CarColorEntity(val id: Long,
                           val hex: String,
                           val nameEn: String,
                           val nameUz: String,
