+++ data/src/main/java/com/badcompany/data/model/CarModelEntity.kt
@@ -3,5 +3,5 @@ package com.badcompany.data.model
 /**
  * Representation for a [CarModelEntity] fetched from an external layer data source
  */
-data class CarModelEntity(val id: Int,
+data class CarModelEntity(val id: Long,
                           val name: String)
