+++ data/src/main/java/com/badcompany/data/model/PhotoEntity.kt
@@ -3,7 +3,7 @@ package com.badcompany.data.model
 /**
  * Representation for a [PhotoEntity] fetched from an external layer data source
  */
-data class PhotoEntity(val id: Int? = null,
+data class PhotoEntity(val id: Long? = null,
                        val name: String? = null,
                        val type: String? = null,
                        val size: Long? = null,
