+++ domain/src/main/java/com/badcompany/domain/domainmodel/CarColorBody.kt
@@ -1,6 +1,6 @@
 package com.badcompany.domain.domainmodel
 
-data class CarColorBody (val id: Int,
+data class CarColorBody (val id: Long,
                          val hex: String,
                          val nameEn: String,
                          val nameUz: String,
