+++ domain/src/main/java/com/badcompany/domain/domainmodel/CarModelBody.kt
@@ -1,4 +1,4 @@
 package com.badcompany.domain.domainmodel
 
-data class CarModelBody (val id: Int,
+data class CarModelBody (val id: Long,
                          val name: String)
