+++ domain/src/main/java/com/badcompany/domain/domainmodel/PhotoBody.kt
@@ -1,6 +1,6 @@
 package com.badcompany.domain.domainmodel
 
-data class PhotoBody (val id: Int? = null,
+data class PhotoBody (val id: Long? = null,
                       val name: String? = null,
                       val type: String? = null,
                       val size: Long? = null,
