+++ remote/src/main/java/com/badcompany/remote/model/PhotoUploadResponse.kt
@@ -7,7 +7,7 @@ data class PhotoUploadResponse(val code: Int? = null,
                                val message: String? = null,
                                val data: PhotoUploadModel? = null)
 
-data class PhotoUploadModel(val id: Int? = null,
+data class PhotoUploadModel(val id: Long? = null,
                             val name: String? = null,
                             val type: String? = null,
                             val size: Long? = null,
